Revert "IOPs are now compliant with -Wdouble-promotion"
This reverts commit f51a497795ea3856c20b68c0d24f5cc230db7e37.

+++ src/develop/develop.c
@@ -382,23 +382,23 @@ void dt_dev_reload_image(dt_develop_t *dev, const uint32_t imgid)
   dt_dev_invalidate(dev); // only invalidate image, preview will follow once it's loaded.
 }
 
-double dt_dev_get_zoom_scale(dt_develop_t *dev, dt_dev_zoom_t zoom, int closeup_factor, int preview)
+float dt_dev_get_zoom_scale(dt_develop_t *dev, dt_dev_zoom_t zoom, int closeup_factor, int preview)
 {
-  double zoom_scale;
+  float zoom_scale;
 
   const float w = preview ? dev->preview_pipe->processed_width  : dev->pipe->processed_width;
   const float h = preview ? dev->preview_pipe->processed_height : dev->pipe->processed_height;
-  const double ps = dev->pipe->backbuf_width ?
-                   dev->pipe->processed_width/dev->preview_pipe->processed_width :
+  const float ps = dev->pipe->backbuf_width ?
+                   dev->pipe->processed_width/(float)dev->preview_pipe->processed_width :
                    dev->preview_pipe->iscale / dev->preview_downsampling;
 
   switch(zoom)
   {
     case DT_ZOOM_FIT:
-      zoom_scale = fmin(dev->width/w, dev->height/h);
+      zoom_scale = fminf(dev->width/w, dev->height/h);
       break;
     case DT_ZOOM_FILL:
-      zoom_scale = fmax(dev->width/w, dev->height/h);
+      zoom_scale = fmaxf(dev->width/w, dev->height/h);
       break;
     case DT_ZOOM_1:
       zoom_scale = closeup_factor;
